+++ src/tools/runPipeline/handler.ts
@@ -7,6 +7,7 @@ import {
 import { runPipelineInputSchema } from './inputSchema.js';
 import mcpErrorOutput from '../../lib/mcpErrorOutput.js';
 import { getCircleCIClient } from '../../clients/client.js';
+import { getAppURL } from '../../clients/circleci/index.js';
 
 export const runPipeline: ToolCallback<{
   params: typeof runPipelineInputSchema;
@@ -23,7 +24,7 @@ export const runPipeline: ToolCallback<{
 
   let projectSlug: string | undefined;
   let branchFromURL: string | undefined;
-
+  const baseURL = getAppURL();
   if (inputProjectSlug) {
     if (!branch) {
       return mcpErrorOutput(
@@ -123,7 +124,7 @@ export const runPipeline: ToolCallback<{
     content: [
       {
         type: 'text',
-        text: `Pipeline run successfully. View it at: https://app.circleci.com/pipelines/${projectSlug}/${runPipelineResponse.number}`,
+        text: `Pipeline run successfully. View it at: ${baseURL}/pipelines/${projectSlug}/${runPipelineResponse.number}`,
       },
     ],
   };